refactor: Create a base popover component to be reused
